* commit 'e1580e948f0c5555fa9cb3357f8266e8f3c6789c':
  JDWP: check thread state on event suspension

+++ jdwp/src/test/java/org/apache/harmony/jpda/tests/jdwp/Events/BreakpointMultipleTest.java
@@ -111,7 +111,7 @@ public class BreakpointMultipleTest extends JDWPEventTestCase {
      */
     private int setBreakpoint(long classID) {
         VmMirror mirror = debuggeeWrapper.vmMirror;
-        return (int) mirror.setBreakpointAtMethodBegin(classID, "breakpointTest");
+        return mirror.setBreakpointAtMethodBegin(classID, "breakpointTest");
     }
 
     /**
